test(runner): output to html.
Implement a simple test runner
that outputs test results to
an HTML page, so that I can
run browser-only javascript
in unit tests, without
transforming back and forth
between Node.js, or relying
on a framework with features
I don't need.

Seriously, why is the tooling
for Javascript such a mess?